minor android refactoring
